Split delete command from trash manager
This separates the required boilerplate of the command from the actual
delete / restore functions.
